refactoring github template copying
