+++ src/commands/checkoutBranch.js
@@ -1,0 +1,6 @@
+import run from '../utils/run';
+
+export default (branchName) => run([
+  'cd tmp/repo',
+  `git checkout -b ${branchName}`,
+]);
