+++ src/commands/copyTemplates.js
@@ -1,0 +1,7 @@
+import run from '../utils/run';
+
+export default () => run([
+  'cd tmp/repo',
+  'mkdir -p .github',
+  'cp -rfv ../../github/templates/* .github',
+]);
