+++ src/githubTemplates.js
@@ -1,8 +1,12 @@
-import fetch from 'node-fetch';
-
 import { validateIsString } from '@pie-dao/utils';
 
-import run from './utils/run';
+import checkoutBranch from './commands/checkoutBranch';
+import cloneRepo from './commands/cloneRepo';
+import commit from './commands/commit';
+import copyTemplates from './commands/copyTemplates';
+import deleteRepo from './commands/deleteRepo';
+import Github from './adapters/Github';
+import pushBranch from './commands/pushBranch';
 
 const repo = process.env.REPO;
 const token = process.env.TOKEN;
@@ -17,35 +21,21 @@ validateIsString(token, {
 
 const main = async () => {
   const branch = `github-templates-${Date.now()}`;
-  const commands = [
-    'mkdir -p tmp',
-    'cd tmp',
-    `git clone git@github.com:${repo} repo`,
-    'cd repo',
-    `git checkout -b ${branch}`,
-    'mkdir -p .github',
-    'cp -rfv ../../github/templates/* .github',
-    'git add .',
-    'git commit -m "update Github templates"',
-    `git push -u origin ${branch}`,
-    'cd ../',
-    'rm -rf repo',
-  ];
 
-  await run(commands);
+  await cloneRepo(repo);
+  await checkoutBranch(branch);
+  await copyTemplates();
+  await commit('update Github templates');
+  await pushBranch(branch);
+  await deleteRepo();
 
-  const url = `https://api.github.com/repos/${repo}/pulls`;
-  const headers = { Authorization: `token ${token}` };
-  const method = 'POST';
-  const body = JSON.stringify({
+  const github = new Github(repo, token);
+
+  console.log(await github.createPullRequest({
+    branch,
+    bodyMessage: 'Updates the Github templates. PR generated by pie-dao/meta repo.',
     title: 'New Github templates',
-    head: branch,
-    base: 'development',
-    body: 'Updates the Github templates. PR generated by pie-dao/meta repo.',
-  });
-
-  const response = await fetch(url, { body, headers, method });
-  console.log(url, await response.json());
+  }));
 };
 
 main();
